client/requirements.txt
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
tqdm==4.66.1